requirements.txt
fastapi==0.101.1
uvicorn[standard]==0.22.0
pandas==2.2.2
numpy==1.26.4
yfinance==0.2.27
python-dotenv==1.0.0
supabase==1.1.0
streamlit==1.25.0
python-multipart==0.0.6
pydantic==1.10.12
